create delete.go
