+++ platform-operator/experimental/module-integration/controllers/verrazzano/delete.go
@@ -1,0 +1,66 @@
+// Copyright (c) 2023, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package verrazzano
+
+import (
+	"context"
+	moduleapi "github.com/verrazzano/verrazzano-modules/module-operator/apis/platform/v1alpha1"
+	"github.com/verrazzano/verrazzano-modules/pkg/controller/result"
+	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
+	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
+	"github.com/verrazzano/verrazzano/platform-operator/constants"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/registry"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+// deleteModules deletes all the modules, optionally only deleting ones that disabled
+func (r Reconciler) deleteModules(log vzlog.VerrazzanoLogger, effectiveCR *vzapi.Verrazzano) result.Result {
+	var reterr error
+	var deletedCount int
+	var moduleCount int
+
+	// If deletion timestamp is non-zero then the VZ CR got deleted
+	fullUninstall := !effectiveCR.GetDeletionTimestamp().IsZero()
+
+	// Delete all modules.  Loop through all the components once even if error occurs.
+	for _, comp := range registry.GetComponents() {
+		if !comp.ShouldUseModule() {
+			continue
+		}
+		moduleCount++
+
+		// If not full uninstall then only disabled components should be uninstalled
+		if !fullUninstall && comp.IsEnabled(effectiveCR) {
+			continue
+		}
+		module := moduleapi.Module{ObjectMeta: metav1.ObjectMeta{
+			Name:      comp.Name(),
+			Namespace: constants.VerrazzanoInstallNamespace,
+		}}
+		err := r.Client.Delete(context.TODO(), &module, &client.DeleteOptions{})
+		if err != nil {
+			if errors.IsNotFound(err) {
+				deletedCount++
+				continue
+			}
+			if !errors.IsConflict(err) {
+				log.Progressf("Failed to delete Component %s, retrying: %v", comp.Name(), err)
+			}
+			reterr = err
+			continue
+		}
+	}
+	if deletedCount != moduleCount {
+		return result.NewResultShortRequeueDelay()
+	}
+
+	// return last error found so that we retry
+	if reterr != nil {
+		return result.NewResultShortRequeueDelayWithError(reterr)
+	}
+	// All modules have been deleted and the Module CRs are gone
+	return result.NewResult()
+}
